requirements.txt
alembic==1.0.0
amqp==2.3.2
aniso8601==3.0.2
beautifulsoup4==4.6.3
billiard==3.5.0.4
blinker==1.4
bs4==0.0.1
celery==4.2.1
click==6.7
Django==1.11
django-braces==1.13.0
django-password-reset==1.0
dominate==2.3.1
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Cache==0.13.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Moment==0.6.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-Session==0.3.1
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
kombu==4.2.1
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
Pillow==5.2.0
python-dateutil==2.7.3
python-editor==1.0.3
python-slugify==1.2.5
pytz==2018.5
redis==2.10.6
six==1.11.0
sorl-thumbnail==12.4.1
SQLAlchemy==1.2.11
Unidecode==1.0.22
vine==1.1.4
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
